Polish build definition
This commit gathers dependency management and common settings in the
parent pom. This has several advantages:

1. dependency management is centralized rather than copy/pasted in
each module
2. Build specific settings are shared (i.e. java version, more on that
below)
3. Some aspects of the build can be inherited rather than duplicated,
such as the `groupId` and the `version`

Please note that `java.version` doesn't do anything if the parent is not
`spring-boot-starter-parent` as it is the one that translates the
java.version property to the right Maven properties. See root pom for
more details.

Here are some best practices that can be translated directly in our
document:

1. A module should not define a version for a dependency. Rather, that
version should be centralized in one place (the parent pom)
2. If a project inherits from a "local" parent, it is better not to
repeat the groupId and version
3. jar packaging is the default in Maven and does not (and should not)
be added
4. Dependencies in the `dependenciesManagement` section should be listed
alphabetically so that it's easier to manage them
5. Properties should be "grouped" and  listed alphabetically so that's
easier to manage them (for instance a group with build related props
and then a group with versions)
6. Add the maven parent as the first entry of the pom so that we can
immediately see what the hierarchy is

(Note that for 6, start.spring.io does not comply with this and I am
guilty of not fixing that)

Of course, all those topics are open for debate!