Application configuration
Never add application.properties in a module. Such file need to be added
only in "app" project. The main reasoning is that Spring Boot will not
merge them (it will take the first that is available on the classpath)
which can lead to unexpected results.

In any case, we should never ever provide such a file, except in samples
of course.