Move auto-configuration test at the appropriate place
Auto-configuration tests should be in the same module as the
auto-configuration itself. We suffixed our test classes with Tests,
rather than Test (again something we shouldn't probably enforce but that
we should apply in our own examples for consistency).

The current test was a bit wrong since the runner did not specify any
property and, as a result, the whole configuration was backing off
anyway.

See the updated tests for more details.