Custom starter base definition
A custom starter should depend on "spring-boot-starter". That way, the
base infrastructure wil be available.

This change actually illustrates the kind of issue that arises if this
isn't done: if a user has to add an additional dependency to be able to
get started with minimum fuss, then something is wrong with the custom
starter definition.

Some starters tend to go further (for instance, starters that are
related to web stuff rely on "spring-boot-starter-web" directly). This
isn't a good idea as the opinions conveyed by a higher-level starter may
be incompatible. A typical example of that is that you want to work with
a reactive web app (webflux starter, instead of web starter).